Struct errors::Wrapper

pub struct Wrapper<T>(pub T);

Wrapper is a very strange type which exists for specialization hacks.

Wrapper has a generic parameter T which means we can implement things for Wrapper<T>, or if we're fun, Wrapper<f32> (for example).

It means our implementations can be specialized for each input type. yet, we can still expose interfaces which bind to From<T>.